Add separate directory for templates
